requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Video/Audio Processing
yt-dlp>=2024.1.0
ffmpeg-python==0.2.0

# AI Services
whisperx>=3.1.0
google-cloud-translate==3.12.1
google-cloud-texttospeech==2.16.3
boto3==1.34.0  # For AWS Polly alternative

# Task Queue
celery[redis]==5.3.4
redis==5.0.1

# Storage
boto3==1.34.0  # For S3-compatible storage (Cloudflare R2)

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Deployment
gunicorn==21.2.0

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1

